requirements.txt
tensorflow
Flask
Pillow

-e .